frontend/requirements.txt
# Core dependencies
pandas>=2.2.3,<3.0
numpy>=1.26.0,<2.0
textstat>=0.7.3,<1.0
tensorflow-cpu>=2.18.0,<3.0
scikit-learn>=1.4.0,<2.0
scipy>=1.13.1,<2.0
spacy>=3.7.4,<4.0
vaderSentiment>=3.3.2,<4.0
streamlit>=1.44.1,<2.0
opencv-python-headless
Pillow
torch
torchvision

en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.0/en_core_web_sm-3.7.0-py3-none-any.whl
